refactor: StepType 임포트 경로를 App에서 constants로 변경
